Keep supervisor within architecture bounds
